mm/mmcg: Fix compiler warning in cgroup_mem_sw_info()
Macro "min()" is call in cgroup_mem_sw_info(), but the types of the two
parameters are inconsistent. So, there are some warnings while compiling
the kernel.
The error is introduced by commit: a4eb616187
The local variable 'limit' and 'memsw_limit' are wrongly declared to be
'unsigned long long'. It should be 'unsigned long'. Fixed it.

+++ mm/memcontrol.c
@@ -5036,7 +5036,7 @@ static void mem_cgroup_bind(struct cgroup_subsys_state *root_css)
 void cgroup_mem_sw_info(struct sysinfo *val, struct mem_cgroup *memcg,
 			long *cached, unsigned long pages[])
 {
-	unsigned long long limit, memsw_limit;
+	unsigned long limit, memsw_limit;
 	u64 memsw_limit_pages;
 	long cache_size, rss_size, swap_size;
 	unsigned int i;
